some change in ReviewRepositoryJDBC

+++ PuffyDonut/src/main/java/repositories/ReviewRepository.java
@@ -6,5 +6,5 @@ import models.Review;
 import java.util.List;
 
 public interface ReviewRepository extends OrmRepository<Review> {
-    List<Review> findAllByUserID(int id);
+    List<Review> findAllByUserID(Long id);
 }
